The function call timeout is measured in seconds and can be a float

+++ src/models/FunctionCall.schema.ts
@@ -5,7 +5,7 @@ const functionCallSchema = z.object({
   function: z.string(),
   args: z.array(z.any()),
   kwargs: z.record(z.string(), z.any()),
-  timeout: z.union([z.number().int().nonnegative(), z.null()]),
+  timeout: z.union([z.number().nonnegative(), z.null()]),
 });
 
 export default functionCallSchema;
